+++ NEWS.md
@@ -1,6 +1,12 @@
 NEWS
 ===========
 
+### raytracing 0.1.1 (Date: 17 dez 2020)
+- Add tryCatch to the ray_path function in order to avoid stopping
+  when the wave does not propagate: [geo(i + 1)] problem
+- Add sep = ";" in write.csv function at ray and ray_source functions
+- Fix typo error in documentation of betaks function, replacing lat by lon
+
 ### raytracing 0.1.0 (Date: 22 out 2020)
 - ray_source returns great circles
 - betaks now returns spatial polygons